As good as it is going to get
ok at this point the code is as done as it is going to get, I've made it reject anything that isn't 0-3999 and alphabet characters, as near as I can tell, but the only problem is it doesn't reject improper roman numerals. We'll get em next time. I still have 15 minutes but I'm not sure how many minutes I spent reading the prompt before I started the timer, so I'm just going to commit and push now, and send.